Implemented template/framework for future section titles
On Library.js, I implemented the section title for PLAYLISTS. This will serve as the template for the other titles.

After lots of trial and error, I've been able to use an SVG and curve the text, to appear similar to our mockups. I also integrated a library called react-curved-text, however, I decided to forego that in favor of using SVGs.