WIP - Home.js scalability improvements
This is an improvement to the very much non-scalable previous Home.js component. Now, we have an improved Home.js UI that scales for a variety of aspect ratios/resolutions, including mobile devices.

There are some bugs that still need to be worked out, but this is a major improvement to what it was previously. Need to make it look better for Car Thing's resolution (800x480).