Slider carousel improvements and minor scalability improvements
Previously, the slider carousel was quite buggy. Now, it's improved. It infinitely loops so that the user doesn't reach a bound to the left or right when their list of recently played music ends. The overall aesthetics and functionality of it have been improved as well.

When viewing the Home page on a mobile device, the album artwork is now centered, and all content on the page is visible. Previously, the album artwork would at times coincide with the nav bar. This no longer is the case.